Complete simpler re-rewite
The focus was on eliminating the linked lists of linked lists and
replacing them with

(a) fixed size array of contexts (can be a service or a driver)
(b) fixed size array of events to be listened for (with pointer back to context)
(c) No more parent/child cruft.   Every driver has one owner.  A service
    can have an owner, but it really doesn't need to do much more than
    send status messages to it

Added the ring buffer functions for all I/O to/from the various subprocesses

The ringbuffer is used to stream I/O to and from the modem handler,
pppd, ssh, vpn driver and the FME

Also added a template (in doc/template.h) containing file header block
and a stub for the fme_notify driver (udp to the FME to indicate
network state)